Check for rev with invalid JSON in CBLRestPusher
It shouldn't be possible for invalid JSON to end up in the database,
but from issue #1179 it seems to have happened. This adds a check for it
because if it isn't caught at this point, it creates trouble a bit
later when the pusher tries to prepare the dictionary to upload.

Fixes #1179 (I think)