More accurate rev pruning in SQLite storage (fixes endless-merge bug)
The quick-and-dirty pruning algorithm for SQLite wasn't good enough. It
would get rid of all non-leaf revisions of a branch, if there was a
much deeper branch as well. This could cause the parent of the short
branch's leaf to be added again if pulled by the replicator, which
effectively re-creates the conflict if the short branch ended in a
deletion.

The code now does the Right Thing by traversing the rev tree and
preserving all revisions that are within maxRevTreeDepth generations of
a leaf.

Unfortunately this is going to slow down insertions somewhat. It could
be optimized because there are some redundant fetches of revs being
done, but honestly I'd rather save the effort for ForestDB.

Fixes #1217