Updated old workaround in SQLiteStorage to work with the real fix
If a revision being PUT turns out to be a duplicate, we should fix the
parent link of the pre-existing row, because it was probably nulled out
when its previous parent got pruned.

The situation is like this: we used to have
   1 <-- 2/
Where 2 is a deletion. Then rev 1 got pruned because the old pruning
algorithm was sloppy and just deleted all non-current revs below a
certain generation count:
         2/
Then 1 got pulled by the replicator because the server is behind and
still has that rev:
   1     2/
Note that there's no link from 2 to 1 now! Instead 1 is current, so it's
a false conflict. Now the app decides to resolve the conflict by
deleting 1. But this creates a revision identical to 2, so SQLite won't
add it. In this situation we used to clear 1's current flag; but that
just makes 1 go away during the pruning, putting us back in the same
state as before. We also need to fix the parent link from 2 to point to
1:
   1 <-- 2/
Now 1 is part of a branch and won't be pruned, but the branch is
deleted so the conflict is gone.

Fixes #1217