Reduce number of revs sent with GET and _bulk_get requests
The pull replicator sends more local revIDs than it needs to, as
context when requesting documents. This wastes bandwidth.

Also, it only sends these when the doc has attachments (the atts_since
parameter.) It should send them always, so SG can intelligently prune
the history it sends in the response. If there are no attachments it
will use a newly-defined 'revs_from' parameter for these instead.

Finally, if there are no known ancestors to send, it now sends a
newly-defined 'revs_limit' parameter whose value is the db's maximum
rev tree depth, to tell the server the max history size to send.

Fixes #1242