Be more careful with attachment lengths, & other improvements
Attempting to resolve the attachment length mismatch reported in #1207.
Still haven't found the cause, but made sure the pusher and listener
set the Content-Length to the actual data length, not the value from the
"length" property which might be inaccurate. Also added a warning when
the two lengths disagree.

Also some cleanup in CBLAttachment, like better checking of the body
object on creation.

For #1207