Fixed: doc._conflicts passed to map function may include wrong revID
When a new revision of a doc isn't the winner (i.e. it's on a shallower
branch), the view indexing code re-indexes the existing winning revision
instead. It should add the new losing revision to the _conflicts array,
but it was getting mixed up and adding the winning rev. Fixed.

Fixes #1252