Don't return an empty "_conflicts" array from REST API
In a document GET with ?conflicts=true, the code to find the conflicts
could filter out all the revisions and end up with an empty array, then
add the _conflicts property anyway.

Turns out most of the callers of -getAllRevisionsOfDocumentID: needed
to filter out deleted revisions, so I added an includeDeleted parameter.
This simplifies the callers a bit and saves some work, as well as
fixing this bug.

Fixes #1260