Don't stop replicator if a bulk-uploaded doc gets an error
If a call to _bulk_docs returns an error for one of the docs (other
than 401 or 403) we propagated that error to the replicator's error
property, which stops the replication. Unfortunately there have been
situations (due to bugs in either CBL or SG) where a specific document
always triggers a 400 error. This doc would then always abort the
replication, preventing any newer revisions from being pushed

Instead, we now just keep going. The affected revision still isn't
removed from the "pending" list, so the checkpoint won't advance past it
and we'll attempt to push it again the next time.

Fixes #1279