Multipart response from ?open_revs shouldn't have "ok":{...} wrapper
The de facto standard for open_revs (as tested with CouchDB 1.6) is for
the individual parts of the multipart response to be just the plain
revision body. It's only in the JSON output format that each item of
the array has an "ok":{...} wrapped around the body.

For couchbaselabs/couchbase-lite-api#155