Newly-pulled docs shouldn't show up in pusher's pendingDocuments
When the pusher is notified of a new revision, but decides to ignore it
because it's an 'echo' from a pull replication (or because it's rejected
by the filter), it should advance the checkpoint past that rev's
sequence.

This is a good optimization in general, to prevent the next push from
re-scanning the same revisions, but more importantly it tells
CBLReplication.pendingDocuments that those revisions aren't pending.

To fix this I did have to change the timing at which the pusher adds
revs to its internal _pendingSequences set, because it needs to happen
at the same time that skipped sequences are being handled.

Fixes #1274