Prevent premature dealloc of CBL_Router
It was possible for a CBL_Router sending an async response to be
dealloced early, since there was nothing keeping a reference to it
during the entire duration of its lifecycle. (I couldn't actually
reproduce this crash, unfortunately.)

To fix this I added a global set sRunningRouters that the router is
added to at the start of its -run method, and removed from at the end
of its -stopNow method.

Fixes #1325