Detect/handle file errors when saving a new attachment
CBLAttachment checks file paths for validity when initialized, but it
doesn't check for read errors when copying the file to the blob-store.
So if the file is moved or deleted before the revision is saved, the
blob isn't created but there's no error. This causes errors later on
when trying to push the revision.

Fixed by adding error handling when reading the file, and did some
refactoring to clean up the code.

Fixes #1180