Disallow rekeying a db if it's open in multiple CBLManagers
We can't rewrite the db if it's open by multiple handles (SQLite or
ForestDB). The proper solution is to close the other handles and then
re-open them [with the new key] afterwards, but that would be too
difficult to do for 1.3. So for now, just detect the situation and
fail with a descriptive error.

Fixes #1233 (we think)