Fixed ability to store db encryption key in (macOS) Keychain
If in the CBL API you specify an encryption key of `@YES` when opening
a database, it'll generate a random key and store it in the Keychain.
Unfortunately the code to generate the key broke a while ago, but the
unit tests wouldn't reveal the problem if they had already been run
before, because the key would already exist.

I fixed this by:
* Correcting the bug in CBLDatabase that prevented the key from being
  generated if an existing one couldn't be found
* Adding a new private method on CBLManager to delete a database's key
  from the Keychain, which is now called before the unit test runs.