Allow SQLite read access concurrently with a write transaction
This is legal since our databases are in WAL mode. It greatly improves
concurrent access -- the ConcurrentWrites unit test (which has two
reader threads and two writers) runs 2.25x faster on my Mac!

I made a slight change to database opening to ensure that the db is
always set to WAL mode, instead of just setting the mode when the db is
first created.

Fixes #1389