Fix for replicator notifications not getting posted
This works around an OS bug where NSNotificationQueue doesn't post its
notifications when called in PostASAP mode while the runloop is handling
a block.

The workaround is to not call CFRunLoopPerform, but instead use our
existing MYPerformOnThread utility, which schedules the block in a
different way that doesn't trigger the bug.

Fixes #1392