Fix Database_Tests.test26_DocumentExpiry failed on devices
The tested failed only when testing with ForestDB. After investigation, the issue occurred only when the Address Sanitizer is ON. After turning the Address  Sanitizer off, the test can pass as expected. To make the test run without a need to turning of the Address Sanitizer, I reduce the number of documents in test from 10000 to 5000.

#1358