Workaround for getting wrong `Server:` response header on first request
It's possible for an initial HTTP request sent by the replicator to hit
some middleware that returns a different Server: header, instead of SG.
In this case the replicator shouldn't hang onto that first header,
since it'll never end up using replicator optimizations like bulk_get.

I changed the logic so that if the currently cached server header is
not from SG, and we receive a different value in a subsequent response,
we switch to the new value (presumably SG.) But once we've received a
response from SG, we keep that.

Also, this commit adds some extra `ChangeTracker` logging -- the full
request headers are now logged, in an attempt to figure out why the
Cookie header isn't getting sent for one developer (which is what
triggers the above situation.)

For #1464