Disable comparison with NSHTTPCookieStorage in unit tests
Since iOS 10.2 and MacOS 10.12.2, the behavior of the Apple NSHTTPCookieStorage has been changed: An NSHTTPCookie with empty path string will not be added to the storage only if the cookie storage contains some cookies in it. Such change makes some checks that compare the result with the Apple cookie storage failed so the comparison with the Apple cookie is disabled until we are sure that the new behavior is correct or is not a bug.