Improved CBLGetHostName on Mac; it was wrong in some network conditions
On my home network with my new AT&T-installed router, gethostname()
returns “jens”, not “jens.local” as expected. Apparently the router’s
DHCP server is telling my computer it has an assigned DNS hostname of
“jens”, which overrides the Sharing system pref setting.

This was breaking the P2P unit test — the URL “https://jens/“ works,
but since the hostname is not in .local our cert-validation code doesn’t
accept the self-signed cert, which causes SSL connections to fail.

Quinn from Apple support says SCDynamicStoreCopyLocalHostName is the
correct call, but it’s not available on iOS so I still have to
conditionalize the code.