Fix for queries being broken immediately after opening a db
If a query ran right after opening a db, before reading any documents,
the SharedKeys instance wouldn’t yet have loaded the keys from the db,
and due to a bug in Fleece, the fl_value() SQL function would fail to
find keys in documents.
Picked up the Fleece fix, and some test changes to help track this down.

(This was a cross-platform bug. Fortunately a cross-platform fix too.)

Fixes #1587