Rev trees are now being pruned (thanks to LiteCore). Added perf test.
Documents were being saved with unbounded max revision tree depth,
which led to massive slowdowns saving a document as the number of revs
increased (not to mention extra disk space.) I wrote a simple perf test
that exposed this.

The problem was that CBL wasn’t setting the maxRevTreeDepth parameter
of the C4DocSaveRequest, and 0 was interpreted by LiteCore as
unbounded. I decided that was bad behavior in LiteCore, and that the
value should be persistently configurable as in 1.x, defaulting to 20.
Implemented that in LiteCore.