Clear _allDocsQuery before closing db, so SQLite db will close
Otherwise LiteCore’s SQLite database won’t close when asked to, because
the query object is holding onto a live sqlite3_stmt.
The db does close later on when the CBLDatabase is dealloced, but the
delayed close causes problems in unit tests because the database dir
can be deleted before the db has closed, which triggers a nasty warning
from SQLite.
I also added a check in CBLTestCase that no C4 objects have been leaked,
and a redundant deletion of the database during teardown to trigger
that warning earlier, during the test that caused the problem.