Clear docReader before retry in CBLBulkDownloader
The crash happened due to the Assert(!_docReader) in the -startParted: delegate called. Based on the log in #1770, the issue happened after a retry.

From the code review, we don’t reset the _docReader so if the retry is successful and if the -startParted: method is called again, the assertion will fail.

Note that I have also reviewed CBLMultipartReader but I couldn’t find a possible way that it would end up calling -startedPart method twice.