Fix log message when suspending the replicator in background
When the app is put into the background, the replicator will start a background task to be able to complete the replication. Once the replication is inactive (e.g. IDLE), the replicator will be automatically suspended. Currently the logging message could be misleading that the replicator cannot run background task. This commit corrected the message.