FTS index should use current locale's language by default
If the caller doesn't specify a language for an FTS index, the docs
say it should use the current locale's language. However, this wasn't
implemented. Added the code for this.

(This required changing the expected number of FTS results in the
TuneMark test, because the search for "rock" wasn't finding the word
"rocks". This is actually how I noticed the bug.)