Sped up array/dictionary conversions and Fleece encoding
In the TuneMark benchmark's import phase, the time was dominated by
adding an NSDictionary to an empty document, and encoding the document
to Fleece. I've sped this up, and some other methods that showed up as
hot-spots.

The biggest problem was the large number of -conformsToProtoco: calls.
This check is very slow. I've replaced those with category methods on
NSObject, which the CBL classes override. These are declared in
CBLData.h.

There are some more changes in CBLDictionary to reduce the amount of
work done in -setDictionary: and -toDictionary.