Fixed: Build script breaks if output dir is an absolute path
generate_release_script_zip.sh would fail if the output path (-o flag)
was an absolute path. The problem was this line (and the three others
like it):

OUTPUT_OBJC_ENTERPRISE_ZIP=`pwd`/$OUTPUT_DIR/couchbase-lite-objc_enterprise$VERSION_SUFFIX.zip

This correctly converts a relative path to absolute, which is necessary
later when the zip file is being built; but if $OUTPUT_DIR is already
an absolute path, it creates a bogus path with the CWD prepended.

I fixed this by making the '*_ZIP' paths relative, but with the correct
"../" prefix to point them to the correct directory.