Add script for stripping frameworks and dSYM files
This is a script to add to the build phase in order to strip non-valid architecture types from the frameworks and dSYM files if added to the built product directory. This is require when archiving the app for submission as currently XCode will not strip non-valid archs (e.g. simulator acrch).

The next step is to figure out where to distribute this file: [1] in release zip or [2] in the framework bundle inself.

#1610