Allow creating Query from JSON, and getting JSON [API CHANGE]
Added to CBLQuery:
  - initWithDatabase:JSONRepresentation:
  @property (nonatomic, readonly) NSData* JSONRepresentation;
(API is provisional. In particular, should there be a factory method
on CBLDatabase, instead of an initializer?)

Restructured the implementation to avoid dependencies on the
query-builder objects (CBLQuerySelectResult, etc.) Instead, the
existing initializer translates to JSON immediately and only stores
JSON. The column names are obtained from a new LiteCore function
(c4query_columnTitle) instead of the builder objects.

Removed some unused code: the properties for the builder objects,
and the -encodeExpressions: method.

couchbaselabs/couchbase-lite-apiv2#212