Hold strong reference to query instance to prevent deallocation from causing crash.
If a `CBLLiveQuery`'s `_query` var in dealloced, this can cause a crash in the `dispatch_after` call because the queue pass in is nil. This can be triggered in a race condition where the live query in dealloced around the same time as a database update that would cause the live query to fire.