Fix custom logger bridging between swift and objc
The original approach that create a Swift custom logger class that implements the private CBLLogger protocol doesnt work as Swift compiler needs to read the CBLLogger definition but as the CBLLogger is private, the private CBLLogger.h is not present. This caused an internal error on XCode.

Instead the fix is using custom logger block to bridge b/w swift and objc custom logger.