Fixed an unused-code warning from Xcode 10.2
Xcode 10.2b1 complains about one line in CBLVersion.m, saying “code
will never be executed” because the result of the strcmp call is
known at compile time, which means the result of the ?: is also
known. (IIRC, the ability to evaluate strcmp et al at compile time
is a new Clang feature.)
Clang suggested wrapping parens around the 0 to mark the expression
as explicitly unused, so I did so.